+++ dspic33c-touch-can-lin-curiosity-touch-canfd.X/mcc_generated_files/system/src/interrupt.c
@@ -15,7 +15,7 @@
 */
 
 /*
-© [2023] Microchip Technology Inc. and its subsidiaries.
+© [2024] Microchip Technology Inc. and its subsidiaries.
 
     Subject to your compliance with these terms, you may use Microchip 
     software and any derivatives exclusively with Microchip products. 
@@ -55,9 +55,9 @@ void INTERRUPT_Initialize(void)
     // Priority: 1
     IPC11bits.DMTIP = 1;
     
-    // CCT4: CCP4 Timer Event
+    // T1: Timer 1
     // Priority: 1
-    IPC10bits.CCT4IP = 1;
+    IPC0bits.T1IP = 1;
     
 }
 
@@ -67,5 +67,5 @@ void INTERRUPT_Deinitialize(void)
     IPC6bits.C1IP = 4;
     IPC1bits.CCT1IP = 4;
     IPC11bits.DMTIP = 4;
-    IPC10bits.CCT4IP = 4;
+    IPC0bits.T1IP = 4;
 }